👏 [Tests] Display customOverlayWidget / customBetweenFabWidget
